# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api SDE for mobile sde",
        "description": "Implementation of Swagger with in Laravel",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://haki44.eu/api",
            "description": "API SDE"
        }
    ],
    "paths": {
...
        "/contact/{id}": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Show one contact",
                "description": "Show one contact",
                "operationId": "ShowContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Contact"
                ],
                "summary": "Update one contact",
                "description": "Update one contact",
                "operationId": "UpdateContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/event/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get one Event",
                "description": "Get one Event",
                "operationId": "ShowEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update one Event",
                "description": "Update one Event",
                "operationId": "UpdateEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete one Event",
                "description": "Delete one Event",
                "operationId": "DeleteEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/period/{id}": {
            "put": {
                "tags": [
                    "Periods"
                ],
                "summary": "Update one period",
                "description": "Update one period",
                "operationId": "UpdatePeriod",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of period",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Periods"
                ],
                "summary": "Delete one period",
                "description": "Delete one period",
                "operationId": "DeletePeriod",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of period",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}